rename: 베이스로 사용되는 Section 컴포넌트와의 혼동을 방지하기 위해 파일명 변경
