+++ src/app/(main)/inquiry/new/_components/Inquiry.tsx
